chore!: extract decoder code
Separation of concerns by moving decoding logic in new class.
