+++ packages/enr/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./constants.js";
 export * from "./creator.js";
+export * from "./decoder.js";
 export * from "./enr.js";
 export * from "./peer_id.js";
 export * from "./waku2_codec.js";
